# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Blog Pessoal",
    "version": "1.0.0",
    "description": "API RESTful completa para um sistema de blog pessoal desenvolvida em Laravel com autenticação JWT como trabalho de conclusão da matéria Sistemas Distribuídos e API da PÓS em FullStack da Unimar.",
    "contact": {
      "name": "Fabiano Souza de Oliveira",
      "email": "fsouzaweb@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autenticação JWT. Use: Bearer {seu_token_jwt}"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Fabiano Souza"},
          "email": {"type": "string", "format": "email", "example": "fabiano@blog.com"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Registrar novo usuário",
        "description": "Cria uma nova conta de usuário no sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "password_confirmation"],
                "properties": {
                  "name": {"type": "string", "example": "Fabiano Souza"},
                  "email": {"type": "string", "format": "email", "example": "fabiano@blog.com"},
                  "password": {"type": "string", "format": "password", "example": "password"},
                  "password_confirmation": {"type": "string", "format": "password", "example": "password"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {"$ref": "#/components/schemas/User"},
                    "token": {"type": "string", "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."}
                  }
                }
              }
            }
          },
          "422": {"description": "Dados de validação inválidos"},
          "500": {"description": "Erro interno do servidor"}
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Fazer login",
        "description": "Autentica um usuário e retorna um token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "fabiano@blog.com"},
                  "password": {"type": "string", "format": "password", "example": "password"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {"$ref": "#/components/schemas/User"},
                    "token": {"type": "string", "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."}
                  }
                }
              }
            }
          },
          "401": {"description": "Credenciais inválidas"}
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Fazer logout",
        "description": "Invalida o token JWT atual do usuário",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "Logout realizado com sucesso"}
                  }
                }
              }
            }
          },
          "401": {"description": "Token inválido ou não fornecido"}
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Renovar token JWT",
        "description": "Gera um novo token JWT usando o token atual",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {"type": "string", "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."}
                  }
                }
              }
            }
          },
          "401": {"description": "Token inválido ou expirado"}
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Autenticação"],
        "summary": "Obter dados do usuário autenticado",
        "description": "Retorna os dados do usuário atualmente autenticado",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dados do usuário obtidos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "401": {"description": "Token inválido ou não fornecido"}
        }
      }
    },
...
  }
}